requirements.txt
alembic==1.13.1
attrs==23.2.0
Automat==22.10.0
blinker==1.7.0
cachetools==5.3.2
camelot-py==0.11.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.0
cssselect==1.2.0
decorator==5.1.1
Elixir==0.7.1
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.0.1
Flask-Cors==4.0.0
flask-marshmallow==1.1.0
Flask-Migrate==4.0.5
flask-sock==0.7.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
javascript==1!1.1.2
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.1.0
Mako==1.3.1
MarkupSafe==2.1.4
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
parsel==1.8.1
pbr==6.0.0
pdfminer.six==20231228
Protego==0.3.0
protobuf==4.25.2
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pygsheets==2.0.6
pyOpenSSL==24.0.0
pyparsing==3.1.1
pypdf==4.0.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
qrcode==7.4.2
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
Scrapy==2.11.0
service-identity==24.1.0
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
SQLAlchemy==2.0.25
sqlalchemy-migrate==0.13.0
sqlparse==0.4.4
tabulate==0.9.0
Tempita==0.5.2
tldextract==5.1.1
tqdm==4.66.1
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
w3lib==2.1.2
Werkzeug==3.0.1
wsproto==1.2.0
xlrd==0.7.1
xlwt==0.7.2
zope.interface==6.1
